api: add inferred_freq flag to .edited point
An infered frequency is calculated on the requested points.
A regular index is created from this frequency and the series index
(if no bound is provided) or from the bounds (if presents).
A special care must be done to have consistent tz-awarness between
the bounds and the series index.
This regular index will present nans when the series has no
associated values. The edition markers of such "holes" are set
as False.